+++ Editor/PackageDependenciesValidation.cs
@@ -12,19 +12,28 @@ namespace UnityEditor.PackageManager.AssetStoreValidation
     class PackageDependenciesValidation : BaseValidation
     {
         static readonly string k_DocsFilePath = "package_dependencies_validation.html";
-        HashSet<string> m_PrecompiledReferences = new HashSet<string>();
-        HashSet<string> m_References = new HashSet<string>();
-        HashSet<string> m_PackagesInUse = new HashSet<string>();
-        Dictionary<string, string> m_RegisteredPackagesDlls = new Dictionary<string, string>();
+        List<ReferenceNameAndPath> m_PrecompiledReferencesNameAndPath = new ();
+        List<ReferenceNameAndPath> m_ReferencesNameAndPath = new ();
+        HashSet<string> m_PackagesInUse = new ();
+        Dictionary<string, string> m_RegisteredPackagesDlls = new ();
         const string k_GuidReferencePrefix = "GUID:";
         const char K_GuidReferenceSeperator = ':';
 
+        
+        internal readonly string k_ErrorSingleMissingDependenciesFromAssembly =
+            "The following package dependency {0} is not used and could be removed from the package manifest. " +
+            ErrorDocumentation.GetLinkMessage(k_DocsFilePath, "unused-package-dependency-in-manifest");
+        
         internal readonly string k_ErrorMissingDependenciesFromAssembly =
-            "The following package dependencies {0} are not used and could be removed from the package manifest. " +
+            "The following package dependencies \n{0}\nare not used and could be removed from the package manifest. " +
             ErrorDocumentation.GetLinkMessage(k_DocsFilePath, "unused-package-dependency-in-manifest");
 
+        internal readonly string k_ErrorMissingSingleDependencyFromManifest =
+            "The following package dependency {0} is missing from the package manifest. " +
+            ErrorDocumentation.GetLinkMessage(k_DocsFilePath, "missing-dependency-in-manifest");
+        
         internal readonly string k_ErrorMissingDependenciesFromManifest =
-            "The following package dependencies {0} \n are missing from the package manifest. " +
+            "The following package dependencies \n{0}\nare missing from the package manifest. " +
             ErrorDocumentation.GetLinkMessage(k_DocsFilePath, "missing-dependency-in-manifest");
         
         internal readonly string k_ErrorMissingPrecompiledFromManifest =
@@ -33,14 +42,14 @@ namespace UnityEditor.PackageManager.AssetStoreValidation
         
         internal readonly string k_ErrorFormatGuidReference =
             "Malformed guid reference {0}. Please validate your assemblies. " +
-            ErrorDocumentation.GetLinkMessage(k_DocsFilePath, "malformed-guid-refence");
+            ErrorDocumentation.GetLinkMessage(k_DocsFilePath, "malformed-guid-reference");
 
         internal readonly string k_ErrorUnknownReference =
-            "Unknown package reference {0}. Please verify your assembly. " +
+            "Unknown package reference of {0} in {1}. Please verify your assembly. " +
             ErrorDocumentation.GetLinkMessage(k_DocsFilePath, "unknown-reference");
 
         internal readonly string k_ErrorUnknownPrecompiledReference =
-            $"The following precompiled reference: {0} could not be found." +
+            $"The following precompiled reference {0} in {1} could not be found. " +
             ErrorDocumentation.GetLinkMessage(k_DocsFilePath, "unknown-precompiled-reference");
 
         internal IPackageInfoHelper m_PackageInfoHelper;
@@ -50,7 +59,7 @@ namespace UnityEditor.PackageManager.AssetStoreValidation
             TestName = "Package Dependencies";
             TestDescription = "Validate that the package dependencies are useful and complete.";
             TestCategory = TestCategory.DataValidation;
-            SupportedValidations = Array.Empty<ValidationType>();
+            SupportedValidations = new[] { ValidationType.AssetStore };
             m_PackageInfoHelper = new PackageInfoHelper();
         }
 
@@ -70,15 +79,18 @@ namespace UnityEditor.PackageManager.AssetStoreValidation
         void ValidateUnusedPackageDependencies()
         {
             var unUsedPackages = new StringBuilder();
+            var unUsedPackagesTracker = new List<string>();
             foreach (var dependency in Context.PublishPackageInfo.dependencies)
             {
-                if (!m_PackagesInUse.Contains(dependency.Key))
-                {
-                    unUsedPackages.AppendLine($"{dependency.Key}@{dependency.Value}");
-                }
+                if(!m_PackagesInUse.Any(i => i.Contains(dependency.Key)))
+                    unUsedPackagesTracker.Add($"{dependency.Key}@{dependency.Value}");
             }
 
-            CheckErrors(unUsedPackages, k_ErrorMissingDependenciesFromAssembly);
+            unUsedPackages.Append(string.Join(",\n", unUsedPackagesTracker));
+            CheckErrors(unUsedPackages,
+                unUsedPackagesTracker.Count > 1
+                    ? k_ErrorMissingDependenciesFromAssembly
+                    : k_ErrorSingleMissingDependenciesFromAssembly);
         }
 
         void CheckErrors(StringBuilder errors, string message)
@@ -93,7 +105,7 @@ namespace UnityEditor.PackageManager.AssetStoreValidation
         {
             var assemblyDefinitions =
                 Directory.EnumerateFiles(Context.PublishPackageInfo.path, "*.asmdef", SearchOption.AllDirectories)
-                    .ToList();
+                    .Where(i => !i.Split(Path.DirectorySeparatorChar).Any(j => (j.EndsWith("~") || j.StartsWith(".")) && !j.Contains("Samples") && j.Equals("cvs", StringComparison.OrdinalIgnoreCase))).ToList();
             var assemblyReferences = Directory
                 .EnumerateFiles(Context.PublishPackageInfo.path, "*.asmref", SearchOption.AllDirectories).ToList();
 
@@ -103,14 +115,13 @@ namespace UnityEditor.PackageManager.AssetStoreValidation
             {
                 var content = File.ReadAllText(assemblyDefinition);
                 var packageParsedMetaData = JObject.Parse(content);
-
-                ExtractReferencesFromAssembly(packageParsedMetaData, m_References, "references");
-                ExtractReferencesFromAssembly(packageParsedMetaData, m_PrecompiledReferences, "precompiledReferences");
+                var path = Path.Combine(Context.PublishPackageInfo.name, assemblyDefinition.Substring(Context.PublishPackageInfo.path.Length + 1));
+                ExtractReferencesFromAssembly(packageParsedMetaData, m_ReferencesNameAndPath, "references", path);
+                ExtractReferencesFromAssembly(packageParsedMetaData, m_PrecompiledReferencesNameAndPath, "precompiledReferences", path);
             }
         }
 
-        void ExtractReferencesFromAssembly(JObject packageParsedMetaData, HashSet<string> collection,
-            string key)
+        void ExtractReferencesFromAssembly(JObject packageParsedMetaData, List<ReferenceNameAndPath> collection, string key, string path)
         {
             if (!packageParsedMetaData.TryGetValue(key, out var packageVersionsAsJson) ||
                 !packageVersionsAsJson.HasValues || !packageVersionsAsJson.Any())
@@ -120,26 +131,23 @@ namespace UnityEditor.PackageManager.AssetStoreValidation
 
             var references = packageVersionsAsJson.Value<JArray>()?.ToObject<List<string>>();
 
-            foreach (var reference in references)
+            foreach (var reference in (references ?? new List<string>()).Where(reference => !string.IsNullOrWhiteSpace(reference)))
             {
-                if (string.IsNullOrWhiteSpace(reference))
-                {
-                    continue;
-                }
-
-                collection.Add(reference);
+                collection.Add(new ReferenceNameAndPath {name = reference, path = path});
             }
         }
 
         void ValidateReferences()
         {
             var missingReferences = new StringBuilder();
+            var missingReferencesTracker = new List<string>();
 
-            foreach (var reference in m_References)
+            foreach (var reference in m_ReferencesNameAndPath)
             {
-                var packageName = reference.StartsWith(k_GuidReferencePrefix)
+                var packageId = reference.name.StartsWith(k_GuidReferencePrefix)
                     ? GetGuidReference(reference)
-                    : GetPackageNameFromReference(reference);
+                    : GetPackageIdFromReference(reference);
+                var packageName = packageId.Split("@")[0];
 
                 if (string.IsNullOrWhiteSpace(packageName) || packageName.Equals(Context.PublishPackageInfo.name))
                 {
@@ -148,70 +156,89 @@ namespace UnityEditor.PackageManager.AssetStoreValidation
 
                 if (!Context.PublishPackageInfo.dependencies.TryGetValue(packageName, out _))
                 {
-                    missingReferences.AppendLine($"{reference}({packageName})");
+                    var dependency = $"{reference.name}({packageId})";
+                    if(!missingReferencesTracker.Contains(dependency))
+                        missingReferencesTracker.Add(dependency);
                 }
                 else
                 {
-                    m_PackagesInUse.Add(packageName);
+                    m_PackagesInUse.Add(packageId);
                 }
             }
-            
-            CheckErrors(missingReferences, k_ErrorMissingDependenciesFromManifest);
+
+            missingReferences.Append(string.Join(",\n", missingReferencesTracker));
+            CheckErrors(missingReferences,
+                missingReferencesTracker.Count > 1
+                    ? k_ErrorMissingDependenciesFromManifest
+                    : k_ErrorMissingSingleDependencyFromManifest);
         }
 
         void ValidatePrecompiledReferences()
         {
             var missingReferences = new StringBuilder();
 
-            foreach (var precompiledReference in m_PrecompiledReferences)
+            foreach (var precompiledReference in m_PrecompiledReferencesNameAndPath)
             {
-                var fileName = Path.GetFileNameWithoutExtension(precompiledReference);
-                if (!m_RegisteredPackagesDlls.TryGetValue(fileName, out var packageName))
+                var fileName = Path.GetFileNameWithoutExtension(precompiledReference.name);
+                if (!m_RegisteredPackagesDlls.TryGetValue(fileName, out var packageId))
                 {
-                    AddError(string.Format(k_ErrorUnknownPrecompiledReference, precompiledReference));
+                    AddError(string.Format(k_ErrorUnknownPrecompiledReference, precompiledReference.name, precompiledReference.path));
                     continue;
                 }
 
-                if (packageName.Equals(Context.PublishPackageInfo.name))
+                var packageName = packageId?.Split("@")[0];
+                if (packageName?.Equals(Context.PublishPackageInfo.name) == true)
                 {
                     continue;
                 }
 
                 // check if the precompiled reference is in the package dependency
-                if (!Context.PublishPackageInfo.dependencies.TryGetValue(packageName, out _))
+                if (!Context.PublishPackageInfo.dependencies.TryGetValue(packageName ?? string.Empty, out _))
                 {
-                    missingReferences.AppendLine($"{precompiledReference}({packageName})");
+                    missingReferences.AppendLine($"{precompiledReference.name}({packageId})");
                 }
                 else
                 {
-                    m_PackagesInUse.Add(packageName);
+                    m_PackagesInUse.Add(packageId);
                 }
             }
 
             CheckErrors(missingReferences, k_ErrorMissingPrecompiledFromManifest);
         }
 
-        string GetGuidReference(string referenceName)
+        string GetGuidReference(ReferenceNameAndPath reference)
         {
-            var referenceItem = referenceName.Split(K_GuidReferenceSeperator);
+            var referenceItem = reference.name.Split(K_GuidReferenceSeperator);
             if (referenceItem.Length != 2)
             {
-                AddError(string.Format(k_ErrorFormatGuidReference, referenceName));
+                AddError(string.Format(k_ErrorFormatGuidReference, reference.name));
                 return string.Empty;
             }
 
-            var packageName = m_PackageInfoHelper.GetPackageNameFromGuid(referenceItem[1]);
-            if (!string.IsNullOrWhiteSpace(packageName)) return packageName;
-            AddError(string.Format(k_ErrorUnknownReference, referenceName));
+            var packageId = m_PackageInfoHelper.GetPackageIdFromGuid(referenceItem[1]);
+            if (!string.IsNullOrWhiteSpace(packageId))
+            {
+                return packageId;
+            }
+            AddError(string.Format(k_ErrorUnknownReference, reference.name, reference.path));
             return string.Empty;
         }
 
-        string GetPackageNameFromReference(string referenceName)
+        string GetPackageIdFromReference(ReferenceNameAndPath reference)
         {
-            var name = m_PackageInfoHelper.GetPackageNameFromReferenceName(referenceName);
-            if (!string.IsNullOrWhiteSpace(name)) return name;
-            AddError(string.Format(k_ErrorUnknownReference, referenceName));
+            var packageId = m_PackageInfoHelper.GetPackageIdFromReferenceName(reference.name);
+            if (!string.IsNullOrWhiteSpace(packageId))
+            {
+                return packageId;
+            }
+            AddError(string.Format(k_ErrorUnknownReference, reference.name, reference.path));
             return string.Empty;
+        }
+        
+        private class ReferenceNameAndPath
+        {
+            public string name;
+            public string path;
         }
     }
 }