+++ Editor/ValidationSuite/ValidationSuite/Utilities/NpmUnusableException.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace UnityEditor.PackageManager.AssetStoreValidation.ValidationSuite
-{
-    internal class NpmUnusableException: Exception
-    {
-        public NpmUnusableException(string disableMessage) : base($"Npm cannot be used because it previously failed with message: {disableMessage}") {}
-    }
-}